requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
sympy>=1.11
django-cors-headers>=4.0
numpy>=1.24